pyproject.toml
[project]
name = "track-time-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.4",
    "djangorestframework>=3.16.0",
    "streamlit>=1.47.0",
    "xlsxwriter>=3.2.5",
]
